+++ srcs/utils/transformations.c
@@ -6,12 +6,11 @@
 /*   By: jkasongo <jkasongo@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/09 14:30:04 by jkasongo          #+#    #+#             */
-/*   Updated: 2022/10/31 17:53:44 by jkasongo         ###   ########.fr       */
+/*   Updated: 2022/11/10 09:29:37 by jkasongo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "transformation.h"
-
 
 t_m4	translate_m(t_v3 with)
 {
@@ -35,17 +34,17 @@ t_m4	scale_m(t_v3 with)
 	return (homothetie);
 }
 
-t_m4	rotation_z(float ang)
+t_m4	rotation_z(double ang)
 {
 	t_m4	rot_matrix;
-	float	cos_val;
-	float	sin_val;
+	double	cos_val;
+	double	sin_val;
 
 	if (ang == 0)
 		return (get_identity_matrix());
 	init_m4(&rot_matrix);
-	cos_val = cosf(ang);
-	sin_val = sinf(ang);
+	cos_val = cos(ang);
+	sin_val = sin(ang);
 	rot_matrix.data[0][0] = cos_val;
 	rot_matrix.data[0][1] = -sin_val;
 	rot_matrix.data[1][0] = sin_val;
@@ -55,17 +54,17 @@ t_m4	rotation_z(float ang)
 	return (rot_matrix);
 }
 
-t_m4	rotation_x(float ang)
+t_m4	rotation_x(double ang)
 {
 	t_m4	rot_matrix;
-	float	cos_val;
-	float	sin_val;
+	double	cos_val;
+	double	sin_val;
 
 	if (ang == 0)
 		return (get_identity_matrix());
 	init_m4(&rot_matrix);
-	cos_val = cosf(ang);
-	sin_val = sinf(ang);
+	cos_val = cos(ang);
+	sin_val = sin(ang);
 	rot_matrix.data[0][0] = 1;
 	rot_matrix.data[1][1] = cos_val;
 	rot_matrix.data[1][2] = -sin_val;
@@ -75,17 +74,17 @@ t_m4	rotation_x(float ang)
 	return (rot_matrix);
 }
 
-t_m4	rotation_y(float ang)
+t_m4	rotation_y(double ang)
 {
 	t_m4	rot_matrix;
-	float	cos_val;
-	float	sin_val;
+	double	cos_val;
+	double	sin_val;
 
 	if (ang == 0)
 		return (get_identity_matrix());
 	init_m4(&rot_matrix);
-	cos_val = cosf(ang);
-	sin_val = sinf(ang);
+	cos_val = cos(ang);
+	sin_val = sin(ang);
 	rot_matrix.data[0][0] = cos_val;
 	rot_matrix.data[0][2] = sin_val;
 	rot_matrix.data[1][1] = 1;
